es esta migrando el proyecto a vitejs con javascript vanilla
