+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>memo-test</title>
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="style.css">
 </head>
 <body>
     <div class="game-wrapper">
@@ -20,6 +20,6 @@
     </div>
 
     
-    <script src="script.js"></script>
+    <script type="module" src="main.js"></script>
 </body>
 </html>
