--- styles.css
@@ -1,121 +1,0 @@
-@font-face {
-    font-family: MoreSugar;
-    src: url('./fonts/MoreSugar-Regular.otf');
-}
-@font-face {
-    font-family: Baloo;
-    src: url('./fonts/Baloo-Regular.ttf');
-}
-
-* {
-    padding: 0;
-    margin: 0;
-    box-sizing: border-box;
-}
-
-body {
-    background-color: #EF4B5A;
-}
-
-.game-wrapper {
-    box-sizing: content-box;
-    width: 26.87em;
-    padding: 2.5em 3em;
-    background-color: white;
-    position: absolute;
-    transform: translate(-50%, -50%);
-    left: 50%;
-    top: 50%;
-    border-radius: 40px;
-    box-shadow: 0 0.9em 2.8em rgba(86, 66, 0, 0.2);
-}
-
-h1 {
-    color: #EF4B5A;
-    text-align: center; 
-    font-family: MoreSugar;
-    font-size: 45px;
-}
-
-.game-stats {
-    color: #073B36;
-    font-family: Baloo;
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
-
-#moves-count,
-#time {
-  margin: 0;
-  padding: 10px; 
-}
-
-.game-board {
-    position: relative;
-    width: 100%;
-    display: grid;
-    gap: 0.6em;
-}
-
-.card-container {
-    position: relative;
-    width: 6.25em;
-    height: 6.25em;
-    cursor: pointer;
-}
-.card-before,
-.card-after {
-    position: absolute;
-    border-radius: 20px;
-    width: 100%;
-    height: 100%;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    border: 2px solid #000000;
-    transition: transform 0.7s ease-out;
-    backface-visibility: hidden;
-}
-
-.card-before {
-    background-color: #9BB69B;
-}
-.card-after {
-    background-color: #9BB69B;
-    transform: rotateY(180deg);
-}
-
-.card-container.flipped .card-before {
-    transform: rotateY(180deg);
-}
-.card-container.flipped .card-after {
-    transform: rotateY(0deg);
-}
-
-.card-container .card-after .image {
-    max-width: 70%;
-    max-height: 70%;
-}
-
-.button-control,
-button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    cursor: pointer;
-}
-
-.button-control {
-    padding: 20px 0px 0px 0;
-}
-
-button {
-    background-color: #EF4B5A;
-    font-family: MoreSugar;
-    font-size: 30px;
-    color: white;
-    border: none;
-    border-radius: 15px;
-    width: 140px;
-}
